split out common biocontrol write serializers
