+++ api-rework/invasives/api/serializers/plant_write_serializers/common/biocontrol_count.py
@@ -1,0 +1,19 @@
+from rest_framework import serializers
+from api.models.activity import (
+    TerrestrialBiocontrolAgentCount,
+    TerrestrialBiocontrolAgentCountExtended,
+)
+
+
+class TerrestrialBiocontrolAgentCountWriteSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = TerrestrialBiocontrolAgentCount
+        fields = ("quantity", "stage")
+
+
+class TerrestrialBiocontrolAgentCountExtendedWriteSerializer(
+    serializers.ModelSerializer
+):
+    class Meta:
+        model = TerrestrialBiocontrolAgentCountExtended
+        fields = ("quantity", "stage", "plant_position", "agent_location")
