+++ api-rework/invasives/api/serializers/plant_write_serializers/common/microsite_conditions.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from api.models.activity import MicrositeCondition
+
+
+class MicrositeConditionsSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = MicrositeCondition
+        fields = ("mesoslope_position", "site_surface_shape")
