+++ api-rework/invasives/api/serializers/plant_write_serializers/common/weather_conditions.py
@@ -1,0 +1,15 @@
+from rest_framework import serializers
+from api.models.activity import WeatherConditions
+
+
+class WeatherConditionsSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = WeatherConditions
+        fields = (
+            "comments",
+            "cloud_cover",
+            "precipitation",
+            "temperature",
+            "wind_direction",
+            "wind_speed_kmh",
+        )
